types: 将类型和接口定义进行拆分
