+++ src/api/login/type.ts
@@ -1,0 +1,14 @@
+export interface ILoginRequestData {
+  /** admin 或 editor */
+  username: "admin" | "editor"
+  /** 密码 */
+  password: string
+  /** 验证码 */
+  code: string
+}
+
+export type LoginCodeResponseData = IApiResponseData<string>
+
+export type LoginResponseData = IApiResponseData<{ token: string }>
+
+export type UserInfoResponseData = IApiResponseData<{ username: string; roles: string[] }>
