+++ src/api/table/index.ts
@@ -1,0 +1,37 @@
+import { request } from "@/utils/service"
+import type * as Table from "./type"
+
+/** 增 */
+export function createTableDataApi(data: Table.ICreateTableRequestData) {
+  return request({
+    url: "table",
+    method: "post",
+    data
+  })
+}
+
+/** 删 */
+export function deleteTableDataApi(id: string) {
+  return request({
+    url: `table/${id}`,
+    method: "delete"
+  })
+}
+
+/** 改 */
+export function updateTableDataApi(data: Table.IUpdateTableRequestData) {
+  return request({
+    url: "table",
+    method: "put",
+    data
+  })
+}
+
+/** 查 */
+export function getTableDataApi(params: Table.IGetTableRequestData) {
+  return request<Table.GetTableResponseData>({
+    url: "table",
+    method: "get",
+    params
+  })
+}
